+++ engine/src/prop/formula.rs
@@ -3,7 +3,7 @@
 use serde::{Deserialize, Serialize};
 
 /// A formula of propositional logic.
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub enum Formula {
     /// An atomic proposition symbol.
     Prop(String),
@@ -22,36 +22,13 @@ pub enum Formula {
     },
 }
 
-/// A formula pattern of propositional logic that can contain meta-variables.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub enum FormulaPattern {
-    /// An atomic proposition symbol.
-    Prop(String),
-
-    /// A unary connective.
-    Unary {
-        connective: UnaryConnective,
-        arg: Box<FormulaPattern>,
-    },
-
-    /// A binary connective.
-    Binary {
-        connective: BinaryConnective,
-        lhs: Box<FormulaPattern>,
-        rhs: Box<FormulaPattern>,
-    },
-
-    /// A meta-variable representing some formula.
-    Meta(String),
-}
-
-#[derive(Debug, Clone, Copy, Serialize, Deserialize)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
 pub enum UnaryConnective {
     /// Negation.
     Not,
 }
 
-#[derive(Debug, Clone, Copy, Serialize, Deserialize)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
 pub enum BinaryConnective {
     /// Conjunction.
     And,
