+++ engine/src/prop/mod.rs
@@ -1,4 +1,3 @@
-pub use formula::{Formula, FormulaPattern};
-
 pub mod formula;
 pub mod parser;
+pub mod pattern;
